+++ xpallets/protocol/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
 name = "xpallet-protocol"
 version = "0.1.0"
-authors = ["ChainX community <https://www.chainx.org>"]
+authors = ["The ChainX Authors"]
 edition = "2018"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "1.3.4", features = ["derive"], default-features = false }
 serde = { version = "1.0", optional = true }
+
 # ChainX primitives
 chainx-primitives = { path = "../../primitives", default-features = false }
